framework/requirements.txt
websockets>=11.0
mindsdb-sdk>=0.8.0
crewai>=0.30.0
crewai-tools>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
asyncio-mqtt>=0.13.0